combined borrow and return into loan
