+++ core/sqldatabase/src/main/java/dk/groupa/sqldatabase/controller/BookController.java
@@ -1,8 +1,7 @@
 package dk.groupa.sqldatabase.controller;
 
 import dk.groupa.sqldatabase.entity.Loan;
-import dk.groupa.sqldatabase.service.BorrowService;
-import dk.groupa.sqldatabase.service.ReturnService;
+import dk.groupa.sqldatabase.service.LoanService;
 import jakarta.validation.constraints.NotNull;
 import org.springframework.web.bind.annotation.*;
 
@@ -10,23 +9,21 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/book")
 public class BookController {
 
-    private final BorrowService borrowService;
-    private final ReturnService returnService;
+    private final LoanService loanService;
 
-    public BookController(BorrowService borrowService, ReturnService returnService) {
-        this.borrowService = borrowService;
-        this.returnService = returnService;
+    public BookController(LoanService loanService) {
+        this.loanService = loanService;
     }
 
     @PostMapping("/{bookId}/borrow")
     public Loan borrowBook(@NotNull @PathVariable("bookId") Integer bookId) {
         int userId = 1;     //TODO: Replace med Mads auth service
-        return borrowService.BorrowBook(bookId, userId);
+        return loanService.BorrowBook(bookId, userId);
     }
 
     @PostMapping("/{bookId}/return")
     public Loan returnBook(@NotNull @PathVariable("bookId") Integer bookId) {
         int userId = 1;     //TODO: Replace med Mads auth service
-        return returnService.ReturnBook(bookId, userId);
+        return loanService.ReturnBook(bookId, userId);
     }
 }
