+++ core/sqldatabase/src/main/java/dk/groupa/sqldatabase/repository/LoanRepository.java
@@ -7,7 +7,10 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BorrowRepository extends JpaRepository<Loan, Long> {
+public interface LoanRepository extends JpaRepository<Loan, Long> {
     @Query(value = "EXEC sp_BorrowBook @user_id = :user_id, @book_id = :book_id", nativeQuery = true)
     Loan borrowBook(@Param("user_id") int user_id, @Param("book_id") int book_id);
+
+    @Query(value = "EXEC sp_ReturnBook @user_id = :user_id, @book_id = :book_id", nativeQuery = true)
+    Loan returnBook(@Param("user_id") int user_id, @Param("book_id") int book_id);
 }
